pyproject.toml
[project]
name = "generative-latent-optimization"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Yus314", email = "shizhaoyoujie@gmail.com" }
]
requires-python = ">=3.10, <3.12"
dependencies = [
    # Core ML framework
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    # Diffusion models and transformers
    "diffusers>=0.20.0",
    "transformers>=4.30.0",
    "accelerate>=0.20.0",
    "safetensors>=0.3.0",
    # Image processing and computer vision
    "pillow>=9.0.0",
    "opencv-python>=4.5.0",
    "scikit-image>=0.20.0",
    "imageio>=2.30.0",
    # Metrics and evaluation
    "lpips>=0.1.4",
    "pytorch-fid>=0.3.0",
    "torchmetrics>=1.0.0",
    # Data processing and visualization
    "numpy>=1.21.0",
    "scipy>=1.10.0",
    "matplotlib>=3.5.0",
    "tqdm>=4.60.0",
    # Configuration and utilities
    "pyyaml>=6.0",
    "requests>=2.25.0",
    # Custom toolkit
    "vae-toolkit @ file:///home/kaki/ghq/github.com/mdipcit/vae-toolkit",
    "h5py>=3.14.0",
    "pytest>=8.4.1",
    "pytest-mock>=3.14.1",
]


[build-system]
requires = ["uv_build>=0.8.6,<0.9.0"]
build-backend = "uv_build"
